Support mega surfaces in Windows, Planes, and Sprites
I'm implementing this via "child" bitmaps, which determine which part of the parent will be visible, manually shrink it if necessary, and send back new values for zoom and offsets.

Window contents and Planes are currently fully functional. Window skins still aren't supported, because it seemed unlikely that would be needed.
Sprites support everything except patterns.