Add a config option to control how the font size is determined.
Some Pokemon Essentials games were explicitly made for mkxp. Text rendered using the old method tends to be smaller than what the new one produces, which means text in those games tends to get cut off with it.

Since Pokemon Essentials uses RGSS1, the setting defaults to the old method for RGSS1, and the new method for RGSS2 and RGSS3.